Use watchable GVRs
Previously, the kubernetes-dependency-watches library would just return
the first GVR matching the GVK. In many cases that was correct, but it
was notably incorrect in the case of template.openshift.io/v1 Templates,
where it would cause the config-policy-controller to try to use the
incorrect "processedtemplates" endpoint.

This updates the version of the kubernetes-dependency-watches library
and adds a test for this situation. Without the correct library version,
the message would mention "processedtemplates" and the test would fail.

Refs:
 - https://issues.redhat.com/browse/ACM-18827

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 30811e4c126660eed86910323e4f2a10432d9078)