Handle more getObject errors
Previously, several invocations were completely ignoring the error from
this function, sometimes commenting that it would be handled later. But
it was not necessarily being handled: in some situations an API error
could cause the controller to act as if a desired resource was missing,
when it was not.

Unfortunately, this error case is hard to exercise in a test.

Refs:
 - https://issues.redhat.com/browse/ACM-21944

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 46e1c76c2386213d88284e693b4340a283b9b199)