Set a maximum compliance message length of 4096
At some point in the past, we had a maximum length of 1024 characters,
but this was insufficient for some policies. Because the messages are
now saved in the status, the overall size needs to be considered again.
This limit should be sufficient.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 3d7596f0ac6ae489ece84b8c906e483e98152532)