fix: Properly handle `objectSelector` with `namespaceSelector`
The `relevantNsNames` variable is a map to a map.
However, in Golang a map is a pointer, so setting
one relevantNsNames[ns] resulted in setting all of
the namespaces to the same value. This instead
instantiates a new value each time.

ref: https://issues.redhat.com/browse/ACM-22676
Signed-off-by: Dale Haiducek <19750917+dhaiducek@users.noreply.github.com>
(cherry picked from commit 2ac5d70aaf1f257e1472fadc64272b018fffe147)