Update client-go to v0.32.9
Previously, if standalone-hub-templating was used and the permissions on
that service account changed (in particular, so that a watched object
could no longer be watched), the retrywatcher would frequently retry and
emit "unknown" errors. Now, the retrywatcher emits a better error, which
the dependency-watches library can use to more intelligently stop the
watch.

Refs:
 - https://issues.redhat.com/browse/ACM-24976

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>