Ensure new namespaces can trigger a reconcile
Previously, if the controller restarted, then policies with an
evaluationInterval would not always start watching the namespaces
specified in their namespaceSelectors. The intention was that new or
modified namespaces could trigger evaluations even outside of the
specified intervals, but this was not working.

Refs:
 - https://issues.redhat.com/browse/ACM-24595

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c6e2628228f2acf5602905fc0d890465288749ae)