Correct compliance on no-op dryrun updates
Previously, the policy status would incorrectly have a noncompliant
event followed immediately by a compliant event when a dryrun update
would not make any changes on that object. This would cause the policy
status to repeatedly update if the policy was reconciling often - for
example if any of its watched objects were updating.

Now only the correct "compliant" event should be emitted in that case.

NOTE: This is a backport, but not just a cherry-pick. This incorporates
some of the changes from these upstream PRs, with updates for tests:
- https://github.com/open-cluster-management-io/config-policy-controller/pull/404
- https://github.com/open-cluster-management-io/config-policy-controller/pull/377

Refs:
 - https://issues.redhat.com/browse/ACM-25742

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>