Record correct OperatorGroup in status
It seems like this was a typo in the original implementation - the
controller correctly updates the resource to match the "merged" version,
but was then using the "desired" version of the resource (which only
specifies what parts the user was interested in, and does not reflect
the full live state) when updating the status of the policy.

The updated section in `musthaveOpGroup` now more closely matches the
analogous section in `musthaveSubscription`.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c3ce2715073c2d55a0caad40a03682000ddbca7e)