Fix operatorpolicy confusing status when invalid
When the subscription section of the OperatorPolicy is invalid, the status might incorrectly indicate that other parts of the policy are invalid. This, combined with the many "... because the policy is invalid" clauses in the overall status, makes it very confusing to read and understand what parts might need to be fixed by the user.
Ref: https://issues.redhat.com/browse/ACM-16781
Signed-off-by: yiraeChristineKim <yikim@redhat.com>
(cherry picked from commit fa7589b5092dead9376d7b827cac623718450222)