Specially handle some OperatorPolicy fields
The optional `config` field in a Subscription, and the optional
`targetNamespaces` and `selector` fields in an OperatorGroup often need
to exactly match what the user has specified, rather than being merged
with what might already be there. That is, when they are specified, it
is usually better to treat them with `mustonlyhave` semantics.

Previously, if a user adjusted the `targetNamespaces` to change it from
one namespace to another (maybe they had a typo originally), the
existing list and the new list would be merged, and the operator would
likely not deploy as intended. Similarly, if a user had a nodeSelector
in their `config`, adjustments would often result in a selector that did
not match any nodes.

Now, when one of those specific fields is set, it is handled better.
When the field is not set, the existing configuration is still used.

Refs:
 - https://issues.redhat.com/browse/ACM-22555

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 208d3b120afd479dfa8605d08d0b6b4518e435df)