Fix duplicated 'no object templates' events
Previously, if no object-templates were present (possibly as a result of
object-templates-raw resolving to be empty), the controller would
incorrectly emit the same event every time the policy was evaluated.
This could cause an accumulation of events on the cluster, and other
issues.

Now, the controller correctly detects the situation, and does not emit
the duplicated events repeatedly.

Refs:
 - https://issues.redhat.com/browse/ACM-27464

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>