Prune some container images in tests
This workflow was failing with 'No space left on device' errors. Having
multiple KinD clusters will use much of the available disk space, so
running the image build while a KinD cluster is running may not be
possible. This commit moves the build earlier, and then cleans up the
build cache to hopefully reclaim enough space for both KinD clusters.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>